refactor(renderer): remove src/webgl directory
